add hook login & register
